Add files via upload
In this short query, I'm trying to create a table with library id and all associated service point ids.
In the inventory_locations table, we have something odd: the same library_id is associated with different names, so I'm assuming these are the primary_service_point names, and not the library names. 
When I run this query, I get one library id associated with three service point ids, but these need to be now broken out in individual columns.